cival-trading-platform/backend/requirements.txt
# MCP Trading Platform - Production Requirements for Railway Deployment
# Optimized for Python 3.11+ compatibility

# Core Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-dotenv>=1.0.0
pydantic==2.10.4
httpx==0.28.1
websockets==14.1

# SSE for real-time updates
sse-starlette>=1.6.0

# Database and ORM
SQLAlchemy>=2.0.0,<2.1.0
psycopg2-binary>=2.9.0

# Cache and Message Queue
redis[hiredis]==5.2.1
aioredis>=2.0.0

# AI and Machine Learning - Production Optimized
openai>=1.0.0
anthropic>=0.25.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Multi-LLM Support
tiktoken>=0.5.0
google-generativeai==0.3.2

# Trading and Financial Data - Production Compatible
yfinance==0.2.50
ccxt>=4.0.0

# Market Data Providers - Multi-source integration
alpha-vantage==2.3.1
polygon-api-client==1.12.3
finnhub-python==2.4.18
pandas-datareader==0.10.0

# Technical Analysis
talib==0.4.25

# Enhanced WebSocket and Real-time
websockets==12.0
python-socketio[asyncio]==5.10.0

# Web3 Integration - Simplified for Production
web3>=6.11.0
eth-account>=0.9.0

# Data Processing and Analysis
requests>=2.28.0
beautifulsoup4>=4.11.0

# Scheduling and Background Tasks
APScheduler>=3.0.0,<4.0.0

# Logging and Monitoring
loguru==0.7.3
structlog==24.4.0
psutil==6.1.0

# Serialization and Data Formats
orjson==3.10.12

# Development and Testing
pytest==8.3.4
pytest-asyncio==0.25.0

# Security and Authentication
passlib[bcrypt]>=1.7.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Async HTTP and Networking
aiohttp>=3.8.0
aiofiles>=23.0.0

# Time and Date Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Crypto and Hashing
cryptography>=41.0.0

# JSON Schema and Validation
jsonschema>=4.17.0

# Memory and AI Agent System
letta>=0.4.0